ai-backend/requirements.txt
fastapi
uvicorn[standard]
ultralytics
tensorflow
pillow
opencv-python-headless
python-multipart